requirements.txt
arrow==1.2.3
asgiref==3.6.0
binaryornot==0.4.4
boto3==1.26.129
botocore==1.29.129
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==2.1.1
Django==4.2.1
django-cors-headers==3.14.0
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
MarkupSafe==2.1.2
mysqlclient==2.1.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.29.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
urllib3==1.26.15
